Extract email sending to a reusable module
