+++ app/assets/stylesheets/searchworks4/search.css
@@ -1,4 +1,6 @@
 .search-query-form {
+  --search-q-border: 1px solid var(--stanford-40-black);
+
   border: 1px solid #ababa9;
   padding: 0.5rem;
   border-radius: 0.25rem;
@@ -35,7 +37,11 @@
 
   .search-q {
     border: 0;
-    border-inline-start: 1px solid var(--stanford-40-black);
+    border-inline-start: var(--search-q-border);
     outline: none;
   }
+}
+
+.database-search-area-bg .search-query-form {
+  --search-q-border: none;
 }
