+++ app/controllers/catalog_controller.rb
@@ -20,6 +20,7 @@ class CatalogController < ApplicationController
   include LocationFacet
 
   include EmailValidation
+  include CatalogEmailSending
 
   include StanfordWorkFacet
 
@@ -510,21 +511,6 @@ class CatalogController < ApplicationController
   end
 
   private
-
-  def send_emails_to_all_recipients(documents)
-    email_params = { message: params[:message], subject: params[:subject], email_from: params[:email_from] }
-    email_addresses.each do |email_address|
-      email_params[:to] = email_address
-      email = if params[:type] == 'full'
-                SearchWorksRecordMailer.full_email_record(documents, email_params, url_options)
-              else
-                SearchWorksRecordMailer.email_record(documents, email_params, url_options)
-              end
-      email.deliver_now
-    end
-
-    flash[:success] = I18n.t('blacklight.email.success')
-  end
 
   def augment_solr_document_json_response(documents)
     documents.map do |document|
