+++ app/views/databases/_popular_databases.html.erb
@@ -5,7 +5,7 @@
           <li class="mb-3">
             <div class="fw-semibold">Academic search premier</div>
             <div>Multidisciplinary database of over 4,650 mostly scholarly publications, many full-text; a great place to start your research.</div>
-            <div><a href="https://research.ebscohost.com">research.ebscohost.com</a> <%= render StanfordOnlyPopoverComponent.new %></div>
+            <div><a href="https://research.ebsco.com/c/qmsjx4?db=aph">research.ebscohost.com</a> <%= render StanfordOnlyPopoverComponent.new %></div>
           </li>
 
           <li class="mb-3">
